Split up Proficiencies component and improved custom proficiency experience
